pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "builder-cli"
version = "0.1.0"
description = "Branch-aware, preset-driven build orchestration tool"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Builder Team", email = "builder@example.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]
dependencies = [
    "PyYAML>=6.0",
]

[project.scripts]
builder = "builder.cli:main"

[tool.setuptools]
package-dir = {"builder" = "builder/src", "core" = "core"}
packages = ["builder", "core"]
